pyproject.toml
[tool.poetry]
name = "penta_sample_fastapi_vue"
version = "0.1.0"
description = ""
authors = ["supplepentan <pentagramog@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.1"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
Jinja2 = "^3.1.2"
Pillow = "^9.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
